Some weird shit going on with the multi-subject training.
Model trains fine with a single subject, 3 subjects also looks
mostly ok, but starts to look a little weird in the attentino maps
also performance isn't great. But even more subjects leads to
serious attention maps issues.

My next step is to change the dataloader to run with mixed batches
and see if that fixes the problem.