Working on mixed dataloader
Still debating whether to do mixed batches or same sub batches
with alternating subjects. Each has advantages and complications.

Will first try mixed batches and the pipeline is easy to setup
but the model may be slow due to need to loop over encoders for
each sample in the batch